+++ iac/lambda.tf
@@ -2,18 +2,19 @@
  * @application Infrastructure-as-Code (IaC)
  * @source lambda.tf
  * @author Bobwares
- * @version 2.0.2
- * @description Per‑verb Lambda functions.
- * @updated 2025-06-21T18:41:22Z
- */
+* @version 2.0.3
+* @description Lambda functions for CRUD actions.
+* @updated 2025-06-25T14:00:08Z
+*/
 
 locals {
   lambda_map = {
-    post   = { handler = "post/index.handler",  policy_json = data.aws_iam_policy_document.ddb_write.json }
-    list   = { handler = "list/index.handler",  policy_json = data.aws_iam_policy_document.ddb_read.json }
-    get    = { handler = "get/index.handler",   policy_json = data.aws_iam_policy_document.ddb_read.json }
-    patch  = { handler = "patch/index.handler", policy_json = data.aws_iam_policy_document.ddb_write.json }
-    delete = { handler = "delete/index.handler", policy_json = data.aws_iam_policy_document.ddb_delete.json }
+    create = { handler = "create.handler", policy_json = data.aws_iam_policy_document.ddb_write.json }
+    list   = { handler = "list.handler", policy_json = data.aws_iam_policy_document.ddb_read.json }
+    get    = { handler = "get.handler", policy_json = data.aws_iam_policy_document.ddb_read.json }
+    update = { handler = "update.handler", policy_json = data.aws_iam_policy_document.ddb_write.json }
+    patch  = { handler = "patch.handler", policy_json = data.aws_iam_policy_document.ddb_write.json }
+    delete = { handler = "delete.handler", policy_json = data.aws_iam_policy_document.ddb_delete.json }
   }
 }
 
@@ -21,7 +22,7 @@ module "lambda" {
   source  = "terraform-aws-modules/lambda/aws"
   version = "7.20.1"
 
-  for_each           = local.lambda_map
+  for_each = local.lambda_map
 
   function_name      = "${local.domain_title}-${each.key}-${var.environment}"
   runtime            = var.lambda_runtime
@@ -31,10 +32,10 @@ module "lambda" {
   policy_json        = each.value.policy_json
 
   environment_variables = {
-    TABLE_NAME    = aws_dynamodb_table.single.name
-    DOMAIN_SCHEMA = jsonencode(local.domain_schema)
+    TABLE_NAME                          = aws_dynamodb_table.single.name
+    DOMAIN_SCHEMA                       = jsonencode(local.domain_schema)
     AWS_NODEJS_CONNECTION_REUSE_ENABLED = 1
-    POWERTOOLS_SERVICE_NAME = local.domain_title
+    POWERTOOLS_SERVICE_NAME             = local.domain_title
   }
 
   tags = merge(var.tags, { Verb = each.key })